Optimized Bezier Computation using parallel SIMD (SSE Only for now)

+++ kernel/inc/preos_renderer.h
@@ -42,6 +42,6 @@ void DrawOsLogo();
 void LineTo(INT64 x0, INT64 y0, INT64 x1, INT64 y1, UINT32 Color);
 
 void CalculateRgba(struct RGBA* Color, struct RGBA* Source);
-extern inline UINT64 GetBezierPoint(double* cordinates, double* beta, UINT8 NumCordinates, double percent);
+extern inline UINT64 GetBezierPoint(float* cordinates, float* beta, UINT8 NumCordinates, float percent);
 
 void TestFill(UINT16 RectX, UINT16 RectY, UINT16 RectWidth, UINT16 RectHeight, UINT32 Color);
